+++ src/parser/shared/metrics/buffApplications.test.ts
