+++ src/parser/shared/metrics/buffUptimeDistinct.test.ts
@@ -1,7 +1,7 @@
 import { ApplyBuffEvent, EventType, RemoveBuffEvent } from 'parser/core/Events';
-import { Info } from 'parser/core/stat';
+import { Info } from 'parser/core/metric';
 
-import buffUptimeTotal from './buffUptimeTotal';
+import buffUptimeDistinct from './buffUptimeDistinct';
 
 const applyBuffEvent = (spellId: number, others?: Partial<ApplyBuffEvent>): ApplyBuffEvent => ({
   type: EventType.ApplyBuff,
@@ -34,7 +34,7 @@ const removeBuffEvent = (spellId: number, others?: Partial<RemoveBuffEvent>): Re
   ...others,
 });
 
-describe('buffUptimeTotal', () => {
+describe('buffUptimeDistinct', () => {
   const info: Info = {
     abilities: [],
     playerId: 1,
@@ -43,23 +43,23 @@ describe('buffUptimeTotal', () => {
   };
 
   it('starts at zero', () => {
-    expect(buffUptimeTotal([], info, 1)).toEqual(0);
+    expect(buffUptimeDistinct([], info, 1)).toEqual(0);
   });
   it('ignores other buffs', () => {
-    expect(buffUptimeTotal([applyBuffEvent(2)], info, 1)).toEqual(0);
+    expect(buffUptimeDistinct([applyBuffEvent(2)], info, 1)).toEqual(0);
   });
   it('calculates the duration for a single application', () => {
     expect(
-      buffUptimeTotal(
+      buffUptimeDistinct(
         [applyBuffEvent(1, { timestamp: 200 }), removeBuffEvent(1, { timestamp: 350 })],
         info,
         1,
       ),
     ).toEqual(150);
   });
-  it('includes stacked time', () => {
+  it('calculates distinct stacked time', () => {
     expect(
-      buffUptimeTotal(
+      buffUptimeDistinct(
         [
           applyBuffEvent(1, { timestamp: 200 }),
           applyBuffEvent(1, { timestamp: 200, targetID: 3 }),
@@ -69,9 +69,9 @@ describe('buffUptimeTotal', () => {
         info,
         1,
       ),
-    ).toEqual(300);
+    ).toEqual(150);
     expect(
-      buffUptimeTotal(
+      buffUptimeDistinct(
         [
           applyBuffEvent(1, { timestamp: 180 }),
           applyBuffEvent(1, { timestamp: 200, targetID: 3 }),
@@ -81,12 +81,12 @@ describe('buffUptimeTotal', () => {
         info,
         1,
       ),
-    ).toEqual(340);
+    ).toEqual(190);
   });
   it('treats the fight end as buff end if removebuff event is missing', () => {
-    expect(buffUptimeTotal([applyBuffEvent(1, { timestamp: 200 })], info, 1)).toEqual(800);
+    expect(buffUptimeDistinct([applyBuffEvent(1, { timestamp: 200 })], info, 1)).toEqual(800);
   });
   it('treats the fight start as buff start if applybuff event is missing', () => {
-    expect(buffUptimeTotal([removeBuffEvent(1, { timestamp: 800 })], info, 1)).toEqual(800);
+    expect(buffUptimeDistinct([removeBuffEvent(1, { timestamp: 800 })], info, 1)).toEqual(800);
   });
 });
