+++ src/parser/shared/metrics/castCount.test.ts
@@ -1,5 +1,5 @@
 import { CastEvent, EventType } from 'parser/core/Events';
-import { Info } from 'parser/core/stat';
+import { Info } from 'parser/core/metric';
 
 import castCount from './castCount';
